refactor(@sounisi5011/encrypted-archive): separate `src/core/utils.ts` contents to multiple files
