+++ packages/encrypted-archive/src/core/utils.ts
@@ -1,39 +1,5 @@
-import type { EncodeStringFn } from './types/builtin';
-import type { AsyncIterableReturn, objectEntries, PartialWithUndefined } from './types/utils';
-
-export function isString(value: unknown): value is string {
-    return typeof value === 'string';
-}
-
-function isFunction(value: unknown): value is () => void {
-    return typeof value === 'function';
-}
-
-export function isNotUndefined<T>(value: T): value is Exclude<T, undefined> {
-    return value !== undefined;
-}
-
-export function isInstance<T extends abstract new (...args: never) => unknown>(instance: T) {
-    return (value: unknown): value is InstanceType<T> => value instanceof instance;
-}
-
-const isNever = (_: unknown): _ is never => false;
-
-export function isOrType<T1, T2, T3 = never, T4 = never>(
-    typeGuard1: (value: unknown) => value is T1,
-    typeGuard2: (value: unknown) => value is T2,
-    typeGuard3: (value: unknown) => value is T3 = isNever,
-    typeGuard4: (value: unknown) => value is T4 = isNever,
-) {
-    return (value: unknown): value is T1 | T2 | T3 | T4 =>
-        typeGuard1(value) || typeGuard2(value) || typeGuard3(value) || typeGuard4(value);
-}
-
-export function isOneArray<T>(value: T[]): value is [T];
-export function isOneArray<T>(value: readonly T[]): value is readonly [T];
-export function isOneArray<T>(value: readonly T[]): value is readonly [T] {
-    return value.length === 1;
-}
+import type { objectEntries, PartialWithUndefined } from './types/utils';
+import { isFunction, isNotUndefined } from './utils/type-check';
 
 export function ifFuncThenExec<TArgs extends unknown[], TResult, TOther>(
     value: ((...args: TArgs) => TResult) | TOther,
@@ -90,59 +56,6 @@ export function normalizeOptions<T extends object>(
                 }, normalizedOptions),
         normalizedOptions,
     );
-}
-
-/**
- * @see https://stackoverflow.com/a/69998555/4907315
- */
-export function uint8arrayConcat(...arrayList: ReadonlyArray<ArrayLike<number>>): Uint8Array {
-    const totalLength = arrayList.map(array => array.length).reduce((a, b) => a + b, 0);
-    const result = new Uint8Array(totalLength);
-
-    let offset = 0;
-    for (const array of arrayList) {
-        result.set(array, offset);
-        offset += array.length;
-    }
-
-    return result;
-}
-
-export function uint8arrayFrom(
-    encodeString: EncodeStringFn,
-    value: string | ArrayBufferView | ArrayBufferLike,
-): Uint8Array {
-    if (typeof value === 'string') return encodeString(value);
-    /**
-     * @see https://github.com/nodejs/node/blob/v12.22.1/lib/zlib.js#L106-L109
-     */
-    if (ArrayBuffer.isView(value)) return new Uint8Array(value.buffer, value.byteOffset, value.byteLength);
-    /**
-     * @see https://github.com/nodejs/node/blob/v12.22.1/lib/zlib.js#L109-L111
-     */
-    return new Uint8Array(value);
-}
-
-export async function* convertIterableValue<T, U>(
-    iterable: Iterable<T> | AsyncIterable<T>,
-    converter: (value: T) => U,
-): AsyncIterableReturn<U, void> {
-    for await (const value of iterable) yield converter(value);
-}
-
-export function number2hex(template: TemplateStringsArray, ...substitutions: number[]): string {
-    return template
-        .map((str, index) => {
-            if (index === 0) return str;
-            const value = substitutions[index - 1];
-            if (typeof value === 'number') {
-                const hexStr = value.toString(16).toUpperCase();
-                const hexLen = hexStr.length + hexStr.length % 2;
-                return `0x${hexStr.padStart(hexLen, '0')}${str}`;
-            }
-            return String(value) + str;
-        })
-        .join('');
 }
 
 export function passThroughString<T, U>(fn: (value: T) => U, value: T): string | U {
